requirements.txt
aiofiles
aiohttp_jinja2
jinja2
aiohttp
xmltodict
pyyaml
ujson
pyhcl
colorama